lsjson: factor internals of lsjson command into operations
